Move OpaqueStyleAndLayoutData to script_layout_interface.
