+++ components/script/lib.rs
@@ -68,6 +68,7 @@ extern crate ref_filter_map;
 extern crate ref_slice;
 extern crate regex;
 extern crate rustc_serialize;
+extern crate script_layout_interface;
 extern crate script_traits;
 extern crate selectors;
 extern crate serde;
