+++ components/script_layout_interface/Cargo.toml
@@ -10,5 +10,7 @@ path = "lib.rs"
 
 [dependencies]
 bitflags = "0.7"
+heapsize = "0.3.0"
+heapsize_plugin = "0.1.2"
 plugins = {path = "../plugins"}
 style = {path = "../style"}
